Reorganize CMake.
