+++ apps/opengl/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(LIBS kimera glfw X11 egl)
+set(LIBS kimera glfw egl)
 
 if(UNIX AND NOT APPLE)
     list(APPEND LIBS dl)
@@ -9,7 +9,6 @@ set (CMAKE_C_STANDARD_LIBRARIES "-lm")
 file(GLOB SOURCE_LIST CONFIGURE_DEPENDS *.c render/glad/*.c)
 
 include(${CMAKE_SOURCE_DIR}/cmake/glfw.cmake)
-include(${CMAKE_SOURCE_DIR}/cmake/X11.cmake)
 include(${CMAKE_SOURCE_DIR}/cmake/egl.cmake)
 
 add_executable(opengl ${SOURCE_LIST})
