+++ apps/tunnel/tunnel.c
@@ -5,12 +5,11 @@
 #include <unistd.h>
 #include <string.h>
 
-#include "kimera/socket.h"
-#include "kimera/encoder.h"
-#include "kimera/decoder.h"
-#include "kimera/resampler.h"
+#include "kimera/transport.h"
+#include "kimera/codec.h"
 #include "kimera/loopback.h"
 #include "kimera/display.h"
+#include "kimera/state.h"
 #include "kimera/client.h"
 
 void receiver(State* state, volatile sig_atomic_t* stop) {
