+++ cmake/dl.cmake
@@ -9,3 +9,5 @@ endif()
 
 target_include_directories(dl INTERFACE ${dl_INCLUDE_DIRS})
 target_link_libraries(dl INTERFACE ${dl_LDFLAGS})
+
+set_property(GLOBAL APPEND PROPERTY GLOBAL_LIB_LIST dl)
