+++ cmake/yaml.cmake
@@ -8,7 +8,10 @@ if(LINUX OR MACOS)
 endif()
 
 if(WINDOWS)
+    message(SEND_ERROR "YAML isn't supported on Windows yet.")
 endif()
 
 target_include_directories(yaml INTERFACE ${YAML_INCLUDE_DIRS})
 target_link_libraries(yaml INTERFACE ${YAML_LDFLAGS})
+
+set_property(GLOBAL APPEND PROPERTY GLOBAL_LIB_LIST yaml)
