+++ include/kimera/client.h
@@ -5,12 +5,17 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdlib.h>
+#include <yaml.h>
 
 #include "kimera/state.h"
-#include "kimera/yaml.h"
 
 volatile sig_atomic_t stop;
 
+//
+// Client Methods
+//
+
+bool kimera_parse_config_file(State* state, char* path);
 void kimera_print_version();
 void kimera_print_state(State*);
 bool kimera_load_state(State*, char*);
