--- include/kimera/tcp_socket.h
@@ -1,17 +1,0 @@
-#ifndef TCP_SOCKET_H
-#define TCP_SOCKET_H
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <unistd.h>
-#include <string.h>
-#include <stdbool.h>
-
-#include "kimera/state.h"
-#include "kimera/socket.h"
-
-bool open_tcp_client(SocketState*, State*);
-bool open_tcp_server(SocketState*, State*);
-void close_tcp(SocketState*);
-
-#endif
