--- include/kimera/udp_socket.h
@@ -1,17 +1,0 @@
-#ifndef UDP_SOCKET_H
-#define UDP_SOCKET_H
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <unistd.h>
-#include <string.h>
-#include <stdbool.h>
-
-#include "kimera/state.h"
-#include "kimera/socket.h"
-
-bool open_udp_client(SocketState*, State*);
-bool open_udp_server(SocketState*, State*);
-void close_udp(SocketState*);
-
-#endif
