--- include/kimera/utils.h
@@ -1,19 +1,0 @@
-#ifndef UTIL_H
-#define UTIL_H
-
-#include <stdint.h>
-#include <stdlib.h>
-
-static inline uint32_t
-buffer_read32be(const uint8_t *buf) {
-    return (buf[3] << 24) | (buf[2] << 16) | (buf[1] << 8) | buf[0];
-}
-
-static inline uint64_t
-buffer_read64be(const uint8_t *buf) {
-    uint32_t msb = buffer_read32be(buf);
-    uint32_t lsb = buffer_read32be(&buf[4]);
-    return ((uint64_t) lsb << 32) | msb;
-}
-
-#endif
