+++ src/codec/CMakeLists.txt
@@ -1,0 +1,9 @@
+if(LINUX OR MACOS OR WINDOWS)
+  list(APPEND LIB_LIST ffmpeg)
+  file(GLOB TMP_LIST CONFIGURE_DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/*.c")
+  list(APPEND SOURCE_LIST ${TMP_LIST})
+endif()
+
+set_property(GLOBAL APPEND PROPERTY GLOBAL_CMAKE_C_FLAGS ${CMAKE_C_FLAGS})
+set_property(GLOBAL APPEND PROPERTY GLOBAL_SOURCE_LIST ${SOURCE_LIST})
+set_property(GLOBAL APPEND PROPERTY GLOBAL_LIB_LIST ${LIB_LIST})
