+++ src/loopback/linux/loopback.c
@@ -1,4 +1,4 @@
-#include "kimera/loopback/linux/loopback.h"
+#include "kimera/loopback/linux.h"
 
 inline bool open_loopback_sink(LoopbackState* loopback, State* state) {
     return init_v4l2_sink(loopback, state);
