+++ src/transport/socket.c
@@ -1,4 +1,4 @@
-#include "kimera/socket.h"
+#include "kimera/transport.h"
 
 inline bool open_socket_client(SocketState* socket, State* state) {
     if (state->source & TCP)
