+++ src/transport/udp_socket.c
@@ -1,4 +1,4 @@
-#include "kimera/udp_socket.h"
+#include "kimera/transport.h"
 
 bool open_udp_client(SocketState* sock_state, State* state) {
     sock_state->server_in = (socket_in*)malloc(sizeof(socket_in));
