+++ src/transport/unix_socket.c
@@ -1,4 +1,4 @@
-#include "kimera/unix_socket.h"
+#include "kimera/transport.h"
 
 bool open_unix_client(SocketState* sock_state, State* state) {
     sock_state->server_un = (socket_un*)malloc(sizeof(socket_un));
